[Add] キャッシュの反映処理を追加
