+++ app/src/marvelous/client/discord/__init__.py
@@ -1,5 +1,3 @@
 from .message import message_gateway
 from .client import bot
 from .start import start
-from .reaction import ReactionEvent
-from .fetch_reaction_event import fetch_reaction_event
