+++ app/src/marvelous/client/discord/events.py
@@ -1,8 +1,8 @@
 from logging import getLogger
 from discord.ext import commands
 import discord
-from ..presentation import check_survival_bonus, add_reaction, remove_reaction
-from .reaction import fetch_reaction_event
+from ..presentation import check_survival_bonus
+from .reaction import set_reaction_state
 
 
 logger = getLogger(__name__)
@@ -20,14 +20,12 @@ async def on_message(message: discord.Message):
 
 @commands.Cog.listener()
 async def on_raw_reaction_add(payload: discord.RawReactionActionEvent):
-    event = await fetch_reaction_event(payload)
-    await add_reaction(event)
+    set_reaction_state(payload, True)
 
 
 @commands.Cog.listener()
 async def on_raw_reaction_remove(payload: discord.RawReactionActionEvent):
-    event = await fetch_reaction_event(payload)
-    await remove_reaction(event)
+    set_reaction_state(payload, False)
 
 
 def setup(bot: commands.Bot):
