+++ app/src/marvelous/client/presentation/__init__.py
@@ -1,6 +1,6 @@
 from .daily_steps import check_reset_daily_steps
 from .survival_bonus import check_reset_survival_bonus, check_survival_bonus
-from .super_marvelous import check_reset_super_marvelous_left, response_super_marvelous
+from .super_marvelous import check_reset_super_marvelous_left
 from .reaction import add_reaction, remove_reaction
 from .user import register_user_implicit, show_status
 from .ranking import show_ranking
