+++ app/src/marvelous/helpers/__init__.py
@@ -1,5 +1,3 @@
 from .number import clamp
 from .iterable import first_match
-from .reaction import get_reaction
-from .user import get_initial_user
 from .time import is_now_time, is_now_weekday
