+++ app/tests/marvelous/models/test_super_marvelous.py
@@ -1,5 +1,6 @@
 import pytest
 import pytest_mock
+from unittest.mock import MagicMock
 from marvelous.models.super_marvelous import *
 from dataclasses import dataclass
 
@@ -76,3 +77,13 @@ def test_cancel(case: SendCase, settings: SuperMarvelousSettings, mocker: pytest
     assert sender.super_marvelous_left == case.before_sender_super_marvelous_left
     assert receiver.point == case.before_receiver_point
     assert super_marvelous.result == case.expected_cancel_result
+
+
+def test_reset_survival_bonus_succeed(mocker: pytest_mock.MockerFixture):
+    """
+    データの更新に成功した場合、結果として成功を表すオブジェクトが返される
+    """
+    count: MagicMock = mocker.Mock()
+    update_func: MagicMock = mocker.patch.object(data_store.users, "reset_super_marvelous_left")
+    reset_super_marvelous_left(count)
+    update_func.assert_called_once_with(count)
