+++ bin/logs.sh
@@ -3,4 +3,4 @@
 cd `dirname $0`
 cd ..
 
-docker-compose -f ./docker-compose.dev.yml logs
+docker-compose logs
