+++ bin/restart.sh
@@ -3,6 +3,5 @@
 cd `dirname $0`
 cd ..
 
-docker-compose -f ./docker-compose.dev.yml down
-#sudo rm -r ./db/data
+./bin/stop.sh
 ./bin/start.sh
