+++ bin/stop.sh
@@ -3,5 +3,5 @@
 cd `dirname $0`
 cd ..
 
-docker-compose -f ./docker-compose.dev.yml down
-#sudo rm -r ./db/data
+docker-compose down
+sudo rm -r ./db/data
