--- docker-compose.dev.yml
@@ -1,30 +1,0 @@
-version: "3.8"
-services:
-  db:
-    container_name: "marvelous_mysql"
-    image: mysql:latest
-    restart: always
-    environment:
-      MYSQL_ROOT_PASSWORD: password
-      MYSQL_DATABASE: marvelous
-      MYSQL_USER: user
-      MYSQL_PASSWORD: password
-      TZ: "Asia/Tokyo"
-    command: mysqld --character-set-server=utf8mb4 --collation-server=utf8mb4_unicode_ci
-    volumes:
-      - ./db/my.cnf:/etc/mysql/conf.d/my.cnf
-      - ./db/initdb.d:/docker-entrypoint-initdb.d
-    ports:
-      - "3306:3306"
-
-  app:
-    build:
-      context: ./app
-      dockerfile: Dockerfile.dev
-    tty: true
-    env_file:
-      - .env
-    depends_on:
-      - db
-    volumes:
-      - ./app:/app
