+++ docker-compose.yml
@@ -12,7 +12,7 @@ services:
       TZ: "Asia/Tokyo"
     command: mysqld --character-set-server=utf8mb4 --collation-server=utf8mb4_unicode_ci
     volumes:
-      - ./db/data:/var/lib/mysql
+      - ./docker/db/data:/var/lib/mysql
       - ./db/my.cnf:/etc/mysql/conf.d/my.cnf
       - ./db/initdb.d:/docker-entrypoint-initdb.d
     ports:
@@ -22,6 +22,15 @@ services:
     build:
       context: ./app
       dockerfile: Dockerfile
+      args:
+        - HTTP_PROXY=$HTTP_PROXY
+        - http_proxy=$http_proxy
+        - HTTPS_PROXY=$HTTPS_PROXY
+        - https_proxy=$https_proxy
+        - FTP_PROXY=$FTP_PROXY
+        - ftp_proxy=$ftp_proxy
+        - NO_PROXY=$NO_PROXY
+        - no_proxy=$no_proxy
     tty: true
     env_file:
       - .env
